requirements.txt
bs4
python-socketio[client]
requests
lxml